src/requirements.txt
loguru==0.7.0
pydantic==2.0.2
flask
pytest
psycopg[binary, pool]==3.1.9
ruff==0.0.282
yoyo-migrations==8.1.0